fastbq/app/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
Jinja2>=3.1.4
google-cloud-bigquery>=3.24.0